🔧 Fix critical deployment issues for Streamlit Cloud
🐛 Issues Fixed:
• Fixed pandas dtype warning by converting to float before FX conversion
• Enhanced agent/__init__.py for proper module imports
• Removed problematic packages.txt (no system deps needed)
• Resolved Python 3.13 compatibility with existing >= versioning

🚀 Deployment Improvements:
• Eliminated FutureWarning spam in logs
• Fixed KeyError: 'agent.planner' import issue
• Removed apt parsing errors from empty packages.txt
• Streamlined dependencies for cloud deployment

✅ Result: Clean deployment logs and stable production app

+++ agent/tools.py
@@ -141,7 +141,8 @@ class FinancialAnalyzer:
             # Convert EUR values to USD
             eur_mask = df_usd['Currency'] == 'EUR'
             if eur_mask.any():
-                df_usd.loc[eur_mask, month] = df_usd.loc[eur_mask, month] * eur_to_usd
+                # Convert to float to avoid dtype warning
+                df_usd.loc[eur_mask, month] = (df_usd.loc[eur_mask, month].astype(float) * eur_to_usd)
                 df_usd.loc[eur_mask, 'Currency'] = 'USD'
         
         return df_usd
